Use the same types in `for` statements.
This removes lots of simple compiler warnings on 'signed' and 'unsigned'
comparisons (reported in #128). Please note that there are two types of
changes. It is because this does not change any function signatures.
  * In most cases: `int` --> `size_t`.
  * In `lookup_table_op.cc`, `string_to_number_op.cc`, and
    `sparse_tensor.h`: `size_t` --> int.

+++ tensorflow/core/framework/tensor.h
@@ -440,7 +440,7 @@ typename TTypes<T, NDIMS>::Tensor Tensor::shaped(
   CHECK_EQ(NDIMS, new_sizes.size());
   int64 new_num_elements = 1;
   Eigen::array<Eigen::DenseIndex, NDIMS> dims;
-  for (int d = 0; d < NDIMS; d++) {
+  for (size_t d = 0; d < NDIMS; d++) {
     new_num_elements *= new_sizes[d];
     dims[d] = new_sizes[d];
   }
@@ -455,7 +455,7 @@ typename TTypes<T, NDIMS>::UnalignedTensor Tensor::unaligned_shaped(
   CHECK_EQ(NDIMS, new_sizes.size());
   int64 new_num_elements = 1;
   Eigen::array<Eigen::DenseIndex, NDIMS> dims;
-  for (int d = 0; d < NDIMS; d++) {
+  for (size_t d = 0; d < NDIMS; d++) {
     new_num_elements *= new_sizes[d];
     dims[d] = new_sizes[d];
   }
@@ -471,7 +471,7 @@ typename TTypes<T, NDIMS>::ConstTensor Tensor::shaped(
   CHECK_EQ(NDIMS, new_sizes.size());
   int64 new_num_elements = 1;
   Eigen::array<Eigen::DenseIndex, NDIMS> dims;
-  for (int d = 0; d < NDIMS; d++) {
+  for (size_t d = 0; d < NDIMS; d++) {
     new_num_elements *= new_sizes[d];
     dims[d] = new_sizes[d];
   }
@@ -486,7 +486,7 @@ typename TTypes<T, NDIMS>::UnalignedConstTensor Tensor::unaligned_shaped(
   CHECK_EQ(NDIMS, new_sizes.size());
   int64 new_num_elements = 1;
   Eigen::array<Eigen::DenseIndex, NDIMS> dims;
-  for (int d = 0; d < NDIMS; d++) {
+  for (size_t d = 0; d < NDIMS; d++) {
     new_num_elements *= new_sizes[d];
     dims[d] = new_sizes[d];
   }
